main/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-PyMongo==2.3.0
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mysqlclient==2.1.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.3.3
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
python-dotenv